Adding support for quota and pam fs limits

+++ node/lib/openshift-origin-node/plugins/unix_user_observer.rb
@@ -30,18 +30,24 @@ module OpenShift
     end
 
     def after_initialize_homedir(user)
+      cmd = "/bin/sh #{File.join('/usr/libexec/openshift/lib', "setup_pam_fs_limits.sh")} #{user.name} #{user.quota_blocks ? user.quota_blocks : ''} #{user.quota_files ? user.quota_files : ''}"
+      out,err,rc = shellCmd(cmd)
+      raise OpenShift::UserCreationException.new("Unable to setup pam/fs limits for #{user.name}") unless rc == 0
     end
 
 
     def before_unix_user_destroy(user)
+      cmd = "/bin/sh #{File.join('/usr/libexec/openshift/lib', "setup_pam_fs_limits.sh")} #{user.name} 0 0 0"
+      out,err,rc = shellCmd(cmd)
+      raise OpenShift::UserCreationException.new("Unable to setup pam/fs/nproc limits for #{user.name}") unless rc == 0
+
       out,err,rc = shellCmd("service cgconfig status > /dev/null")
       if rc == 0
-        shellCmd("/usr/bin/oo-admin-ctl-cgroups stopuser #{user.name} > /dev/null")
+        shellCmd("/usr/bin/oo-admin-ctl-cgroups freezeuser #{user.name} > /dev/null") if rc == 0
       end
 
       last_access_dir = OpenShift::Config.instance.get("LAST_ACCESS_DIR")
       shellCmd("rm -f #{last_access_dir}/#{user.name} > /dev/null")
-
     end
 
     def before_initialize_openshift_port_proxy(user)
@@ -51,6 +57,13 @@ module OpenShift
     end
 
     def after_unix_user_destroy(user)
+      out,err,rc = shellCmd("service cgconfig status > /dev/null")
+      shellCmd("/usr/bin/oo-admin-ctl-cgroup thawuser #{user.name} > /dev/null") if rc == 0
+      shellCmd("/usr/bin/oo-admin-ctl-cgroup stopuser #{user.name} > /dev/null") if rc == 0
+
+      cmd = "/bin/sh #{File.join("/usr/libexec/openshift/lib", "teardown_pam_fs_limits.sh")} #{user.name}"
+      out,err,rc = shellCmd(cmd)
+      raise OpenShift::UserCreationException.new("Unable to teardown pam/fs/nproc limits for #{user.name}") unless rc == 0
     end
 
     def before_add_ssh_key(user,key)
